RC4终极验收修复

+++ qgc-server/qgc-content/src/main/java/com/qiongguichou/content/entity/AdStat.java
@@ -17,9 +17,9 @@ public class AdStat {
 
     private LocalDate statDate;
 
-    private Integer impressionCount;
+    private Long impressionCount;
 
-    private Integer clickCount;
+    private Long clickCount;
 
     @TableField(fill = FieldFill.INSERT)
     private LocalDateTime createTime;
